Fixed node code so that build scripts work.
