+++ src/node/avr_hardware.h
@@ -244,95 +244,3 @@ public:
         eepromWriteWord(EEPROM_ADRW_EXITAT, rssi);
     }
 };
-
-AvrHardware defaultHardware;
-Hardware *hardware = &defaultHardware;
-
-static Message i2cMessage(RssiReceivers::rssiRxs, hardware);
-static Message serialMessage(RssiReceivers::rssiRxs, hardware);
-
-// Function called by twi interrupt service when master sends information to the slave
-// or when master sets up a specific read request
-void i2cReceive(int byteCount)
-{  // Number of bytes in rx buffer
-   // If byteCount is zero, the master only checked for presence of the slave device, no response necessary
-    if (byteCount == 0)
-    {
-        LOG_ERROR("no bytes to receive?");
-        return;
-    }
-
-    if (byteCount != Wire.available())
-    {
-        LOG_ERROR("rx byte count and wire available don't agree");
-    }
-
-    i2cMessage.command = Wire.read();  // The first byte sent is a command byte
-
-    if (i2cMessage.command > 0x50)
-    {  // Commands > 0x50 are writes TO this slave
-        byte expectedSize = i2cMessage.getPayloadSize();
-        if (expectedSize > 0 && i2cReadAndValidateIoBuffer(expectedSize))
-        {
-            i2cMessage.handleWriteCommand(false);
-        }
-        i2cMessage.buffer.size = 0;
-    }
-    else
-    {  // Otherwise this is a request FROM this device
-        if (Wire.available())
-        {  // There shouldn't be any data present on the line for a read request
-            LOG_ERROR("Wire.available() on a read request.", ioCommand, HEX);
-            while (Wire.available())
-            {
-                Wire.read();
-            }
-        }
-    }
-}
-
-bool i2cReadAndValidateIoBuffer(byte expectedSize)
-{
-    uint8_t checksum;
-
-    for (i2cMessage.buffer.size = 0; i2cMessage.buffer.size < expectedSize + 1;
-            i2cMessage.buffer.size++)
-    {
-        if (!Wire.available())
-        {
-            return false;
-        }
-        i2cMessage.buffer.data[i2cMessage.buffer.size] = Wire.read();
-    }
-
-    checksum = i2cMessage.buffer.calculateChecksum(expectedSize);
-
-    if (i2cMessage.buffer.data[i2cMessage.buffer.size-1] == checksum)
-    {
-        return true;
-    }
-    else
-    {
-        LOG_ERROR("Invalid checksum", checksum);
-        return false;
-    }
-}
-
-// Function called by twi interrupt service when a request for data is received
-// No parameters and no returns
-// A transmit buffer (ioBuffer) is populated with the data before sending.
-void i2cTransmit()
-{
-    i2cMessage.handleReadCommand(false);
-
-    if (i2cMessage.buffer.size > 0)
-    {  // If there is pending data, send it
-        Wire.write((byte *)i2cMessage.buffer.data, i2cMessage.buffer.size);
-        i2cMessage.buffer.size = 0;
-    }
-}
-
-void serialEvent()
-{
-    handleStreamEvent(Serial, serialMessage);
-}
