+++ src/node/stm32_hardware.cpp
@@ -1,0 +1,14 @@
+#include "config.h"
+#if STM32_MODE_FLAG && !defined(__TEST__)
+#include "stm32_hardware.h"
+
+Stm32Hardware defaultHardware;
+Hardware *hardware = &defaultHardware;
+
+static Message serialMessage(RssiReceivers::rssiRxs, hardware);
+
+void serialEvent()
+{
+    handleStreamEvent(SERIALCOM, serialMessage);
+}
+#endif
